// src/shared/data/generated-municipalities.ts
export const generatedPrefectures: PrefectureRecord[] = [
  { code: '01', name: '北海道', kana: 'ほっかいどう' },
  { code: '02', name: '青森県', kana: 'あおもりけん' },
  { code: '03', name: '岩手県', kana: 'いわてけん' },
  { code: '04', name: '宮城県', kana: 'みやぎけん' },
  { code: '05', name: '秋田県', kana: 'あきたけん' },
  { code: '06', name: '山形県', kana: 'やまがたけん' },
  { code: '07', name: '福島県', kana: 'ふくしまけん' },
  { code: '08', name: '茨城県', kana: 'いばらきけん' },
  { code: '09', name: '栃木県', kana: 'とちぎけん' },
  { code: '10', name: '群馬県', kana: 'ぐんまけん' },
  { code: '11', name: '埼玉県', kana: 'さいたまけん' },
  { code: '12', name: '千葉県', kana: 'ちばけん' },
  { code: '13', name: '東京都', kana: 'とうきょうと' },
  { code: '14', name: '神奈川県', kana: 'かながわけん' },
  { code: '15', name: '新潟県', kana: 'にいがたけん' },
  { code: '16', name: '富山県', kana: 'とやまけん' },
  { code: '17', name: '石川県', kana: 'いしかわけん' },
  { code: '18', name: '福井県', kana: 'ふくいけん' },
  { code: '19', name: '山梨県', kana: 'やまなしけん' },
  { code: '20', name: '長野県', kana: 'ながのけん' },
  { code: '21', name: '岐阜県', kana: 'ぎふけん' },
  { code: '22', name: '静岡県', kana: 'しずおかけん' },
  { code: '23', name: '愛知県', kana: 'あいちけん' },
  { code: '24', name: '三重県', kana: 'みえけん' },
  { code: '25', name: '滋賀県', kana: 'しがけん' },
  { code: '26', name: '京都府', kana: 'きょうとふ' },
  { code: '27', name: '大阪府', kana: 'おおさかふ' },
  { code: '28', name: '兵庫県', kana: 'ひょうごけん' },
  { code: '29', name: '奈良県', kana: 'ならけん' },
  { code: '30', name: '和歌山県', kana: 'わかやまけん' },
  { code: '31', name: '鳥取県', kana: 'とっとりけん' },
  { code: '32', name: '島根県', kana: 'しまねけん' },
  { code: '33', name: '岡山県', kana: 'おかやまけん' },
  { code: '34', name: '広島県', kana: 'ひろしまけん' },
  { code: '35', name: '山口県', kana: 'やまぐちけん' },
  { code: '36', name: '徳島県', kana: 'とくしまけん' },
  { code: '37', name: '香川県', kana: 'かがわけん' },
  { code: '38', name: '愛媛県', kana: 'えひめけん' },
  { code: '39', name: '高知県', kana: 'こうちけん' },
  { code: '40', name: '福岡県', kana: 'ふくおかけん' },
  { code: '41', name: '佐賀県', kana: 'さがけん' },
  { code: '42', name: '長崎県', kana: 'ながさきけん' },
  { code: '43', name: '熊本県', kana: 'くまもとけん' },
  { code: '44', name: '大分県', kana: 'おおいたけん' },
  { code: '45', name: '宮崎県', kana: 'みやざきけん' },
  { code: '46', name: '鹿児島県', kana: 'かごしまけん' },
  { code: '47', name: '沖縄県', kana: 'おきなわけん' },
]